+++ examples/config/config_fabric_info.yaml
