+++ lib/ndfc_python/validators/reachability.py
@@ -1,0 +1,25 @@
+from ipaddress import IPv4Address
+from typing import List
+
+from pydantic import BaseModel
+
+
+class ReachabilityConfig(BaseModel):
+    """
+    # Summary
+
+    Base validator for Reachability arguments
+    """
+
+    fabric_name: str
+    seed_ip: IPv4Address
+
+
+class ReachabilityConfigValidator(BaseModel):
+    """
+    # Summary
+
+    config is a list of ReachabilityConfig
+    """
+
+    config: List[ReachabilityConfig]
